open_midas/requirements.txt
torch
torchvision
imgaug
matplotlib
numpy
seaborn
IPython
Pillow
tqdm
scikit-learn
networkx
opencv-python
catboost